# platform-service/api-docs.yaml
openapi: 3.0.3
info:
  title: Network Measurement Platform - API
  description: |-
    This is the REST API for our application server based on the OpenAPI 3.0 specification. The API is subject to change during the development process.
    
    Useful links:
    - [GitLab Repository](https://gitlab.lrz.de/cm/2023-karthiknarumanchi-mt)
    - [Wiki on GitLab](https://gitlab.lrz.de/cm/2023-karthiknarumanchi-mt/-/wikis/home)
  version: 0.6.0
  contact:
    email: karthik.narumanchi@tum.de
servers:
  - url: localhost:5000
    description: Development server
...
paths:
...
  /api/results/{measurementId}:
    get:
      tags:
      - Result
      summary: Get result for a measurement by the measurement's id
      operationId: getResultByMeasurementId
      parameters:
      - name: measurementId
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Result not found
        "403":
          description: Forbidden
        "200":
          description: OK
          content:
            application/octet-stream: {}
      security:
      - Bearer_Authentication: []
...
components:
...
  securitySchemes:
    Bearer_Authentication:
      type: http
      description: A JWT token is required to access this API. JWT token can be obtained
        from the auth/request API.
      scheme: bearer
      bearerFormat: JWT